Pet 156 feat : 사용자 탈퇴 및 페밀리 탈퇴 시 검증 로직 추가

+++ Domain-Module/Todo-Module/Todo-Application/src/main/java/com/pawith/todoapplication/handler/UserAccountDeleteOnTodoHandler.java
@@ -1,9 +1,11 @@
 package com.pawith.todoapplication.handler;
 
 import com.pawith.commonmodule.event.UserAccountDeleteEvent;
+import com.pawith.tododomain.entity.Register;
 import com.pawith.tododomain.service.AssignDeleteService;
 import com.pawith.tododomain.service.RegisterDeleteService;
 import com.pawith.tododomain.service.RegisterQueryService;
+import com.pawith.tododomain.service.RegisterValidateService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
@@ -11,6 +13,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Component
@@ -19,12 +22,16 @@ import java.util.List;
 public class UserAccountDeleteOnTodoHandler {
     private final RegisterQueryService registerQueryService;
     private final RegisterDeleteService registerDeleteService;
+    private final RegisterValidateService registerValidateService;
     private final AssignDeleteService assignDeleteService;
 
     @EventListener
-    public void deleteUserInfo(UserAccountDeleteEvent userAccountDeleteEvent){
-        final List<Long> registerIds = registerQueryService.findRegisterIdsByUserId(userAccountDeleteEvent.getUserId());
-        log.info("registerIds: {}", registerIds);
+    public void deleteUserInfo(UserAccountDeleteEvent userAccountDeleteEvent) {
+        final List<Register> registers = registerQueryService.findAllRegistersByUserId(userAccountDeleteEvent.getUserId());
+        registerValidateService.validatePresidentRegisterListDeletable(registers);
+        final List<Long> registerIds = registers.stream()
+            .map(Register::getId)
+            .collect(Collectors.toList());
         assignDeleteService.deleteAssignByRegisterId(registerIds);
         registerDeleteService.deleteRegisterByRegisterIds(registerIds);
     }
